+++ src/main/java/org/stellar/sdk/TrustLineAsset.java
@@ -1,180 +1,103 @@
 package org.stellar.sdk;
 
-import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NonNull;
+import org.jetbrains.annotations.Nullable;
+import org.stellar.sdk.xdr.AssetType;
 
 /**
- * TrustLineAsset class.
- *
- * @see <a
- *     href="https://developers.stellar.org/docs/fundamentals-and-concepts/stellar-data-structures/assets"
- *     target="_blank">Assets</a>
+ * Represents a trustline asset in the Stellar network. A trustline is a declaration that an account
+ * trusts an issuer of an asset up to a certain limit. This class can represent both regular assets
+ * and liquidity pool shares.
  */
-public abstract class TrustLineAsset implements Comparable<TrustLineAsset> {
-  TrustLineAsset() {}
+@EqualsAndHashCode
+@Getter
+public class TrustLineAsset {
+  /** The type of the asset. */
+  @NonNull private final AssetType assetType;
 
   /**
-   * Parses an asset string and returns the equivalent TrustLineAsset instance. The asset string is
-   * expected to either be "native" or a string of the form "CODE:ISSUER"
+   * The asset for which the trustline is established. This is null if the trustline is for a
+   * liquidity pool share.
    *
-   * @param canonicalForm Canonical string representation of an asset
+   * <p>If assetType is one of {@link AssetType#ASSET_TYPE_NATIVE}, {@link
+   * AssetType#ASSET_TYPE_CREDIT_ALPHANUM4} or {@link AssetType#ASSET_TYPE_CREDIT_ALPHANUM12} then
+   * this field will be set.
    */
-  public static TrustLineAsset create(String canonicalForm) {
-    return new Wrapper(Asset.create(canonicalForm));
+  @Nullable private final Asset asset;
+
+  /**
+   * The ID of the liquidity pool for which the trustline is established. This is null if the
+   * trustline is for a regular asset.
+   *
+   * <p>If assetType is {@link AssetType#ASSET_TYPE_POOL_SHARE} then this field will be set.
+   */
+  @Nullable private final LiquidityPoolID liquidityPoolId;
+
+  /**
+   * Creates a TrustLineAsset for a regular asset.
+   *
+   * @param asset The asset for which the trustline is created.
+   */
+  public TrustLineAsset(@NonNull Asset asset) {
+    this.assetType = asset.getType();
+    this.asset = asset;
+    this.liquidityPoolId = null;
   }
 
   /**
-   * Parses type, code, issuer and returns the equivalent TrustLineAsset instance.
+   * Creates a TrustLineAsset for a liquidity pool share.
    *
-   * @param type the asset type
-   * @param code the asset code
-   * @param issuer the assset issuer
-   * @return TrustLineAsset
+   * @param liquidityPoolId The ID of the liquidity pool.
    */
-  public static TrustLineAsset create(String type, String code, String issuer) {
-    return new Wrapper(Asset.create(type, code, issuer));
+  public TrustLineAsset(@NonNull LiquidityPoolID liquidityPoolId) {
+    this.assetType = AssetType.ASSET_TYPE_POOL_SHARE;
+    this.asset = null;
+    this.liquidityPoolId = liquidityPoolId;
   }
 
   /**
-   * Converts Asset to TrustLineAsset
+   * Converts this TrustLineAsset to its XDR representation.
    *
-   * @param asset the Asset
-   * @return TrustLineAsset
+   * @return The XDR representation of this TrustLineAsset.
    */
-  public static TrustLineAsset create(Asset asset) {
-    return new Wrapper(asset);
-  }
-
-  /**
-   * Creates a TrustLineAsset from LiquidityPoolParameters
-   *
-   * @param params the LiquidityPoolParameters
-   * @return TrustLineAsset
-   */
-  public static TrustLineAsset create(LiquidityPoolParameters params) {
-    return new LiquidityPoolShareTrustLineAsset(params);
-  }
-
-  /**
-   * Creates a TrustLineAsset from LiquidityPoolID
-   *
-   * @param id the LiquidityPoolID
-   * @return TrustLineAsset
-   */
-  public static TrustLineAsset create(LiquidityPoolID id) {
-    return new LiquidityPoolShareTrustLineAsset(id);
-  }
-
-  /**
-   * Creates a TrustLineAsset from ChangeTrustAsset
-   *
-   * @param wrapper the ChangeTrustAsset wrapper
-   * @return TrustLineAsset
-   */
-  public static TrustLineAsset create(ChangeTrustAsset.Wrapper wrapper) {
-    return new Wrapper(wrapper.getAsset());
-  }
-
-  /**
-   * Create TrustLineAsset from LiquidityPoolShareChangeTrustAsset
-   *
-   * @param share the LiquidityPoolShareChangeTrustAsset
-   * @return TrustLineAsset
-   */
-  public static TrustLineAsset create(LiquidityPoolShareChangeTrustAsset share) {
-    return new LiquidityPoolShareTrustLineAsset(share.getLiquidityPoolParams());
-  }
-
-  /**
-   * Creates TrustLineAsset based on a <code>code</code> length and issuer only
-   *
-   * @param code the TrustLineAsset code
-   * @param issuer the TrustLineAsset issuer
-   */
-  public static TrustLineAsset createNonNativeAsset(String code, String issuer) {
-    return create(Asset.create(null, code, issuer));
-  }
-
-  /**
-   * Generates TrustLineAsset object from a given XDR object
-   *
-   * @param xdr XDR object
-   */
-  public static TrustLineAsset fromXdr(org.stellar.sdk.xdr.TrustLineAsset xdr) {
-    // TODO: Figure out how we can re-use Asset.fromXdr here
-    String accountId;
-    switch (xdr.getDiscriminant()) {
-      case ASSET_TYPE_NATIVE:
-        return TrustLineAsset.create(new AssetTypeNative());
-      case ASSET_TYPE_CREDIT_ALPHANUM4:
-        String assetCode4 =
-            Util.paddedByteArrayToString(xdr.getAlphaNum4().getAssetCode().getAssetCode4());
-        accountId = StrKey.encodeEd25519PublicKey(xdr.getAlphaNum4().getIssuer());
-        return TrustLineAsset.create(new AssetTypeCreditAlphaNum4(assetCode4, accountId));
-      case ASSET_TYPE_CREDIT_ALPHANUM12:
-        String assetCode12 =
-            Util.paddedByteArrayToString(xdr.getAlphaNum12().getAssetCode().getAssetCode12());
-        accountId = StrKey.encodeEd25519PublicKey(xdr.getAlphaNum12().getIssuer());
-        return TrustLineAsset.create(new AssetTypeCreditAlphaNum12(assetCode12, accountId));
-      case ASSET_TYPE_POOL_SHARE:
-        return new LiquidityPoolShareTrustLineAsset(
-            LiquidityPoolID.fromXdr(xdr.getLiquidityPoolID()));
-      default:
-        throw new IllegalArgumentException("Unknown asset type " + xdr.getDiscriminant());
-    }
-  }
-
-  @Override
-  public abstract boolean equals(Object object);
-
-  @Override
-  public abstract int compareTo(@NonNull TrustLineAsset other);
-
-  /**
-   * Get the asset type
-   *
-   * @return the asset type
-   */
-  public abstract String getType();
-
-  /**
-   * Generates XDR object from a given TrustLineAsset object
-   *
-   * @return xdr model
-   */
-  public abstract org.stellar.sdk.xdr.TrustLineAsset toXdr();
-
-  @Getter
-  @EqualsAndHashCode(callSuper = false)
-  @AllArgsConstructor
-  public static final class Wrapper extends TrustLineAsset {
-    private final Asset asset;
-
-    @Override
-    public String getType() {
-      return asset.getType();
-    }
-
-    @Override
-    public int compareTo(@NonNull TrustLineAsset other) {
-      if ("pool_share".equals(other.getType())) {
-        return -1;
-      }
-      return this.getAsset().compareTo(((TrustLineAsset.Wrapper) other).getAsset());
-    }
-
-    @Override
-    public org.stellar.sdk.xdr.TrustLineAsset toXdr() {
-      org.stellar.sdk.xdr.TrustLineAsset xdr = new org.stellar.sdk.xdr.TrustLineAsset();
-
+  public org.stellar.sdk.xdr.TrustLineAsset toXdr() {
+    org.stellar.sdk.xdr.TrustLineAsset xdr = new org.stellar.sdk.xdr.TrustLineAsset();
+    if (asset != null) {
       org.stellar.sdk.xdr.Asset assetXdr = asset.toXdr();
       xdr.setDiscriminant(assetXdr.getDiscriminant());
       xdr.setAlphaNum4(assetXdr.getAlphaNum4());
       xdr.setAlphaNum12(assetXdr.getAlphaNum12());
+    }
+    if (liquidityPoolId != null) {
+      xdr.setDiscriminant(AssetType.ASSET_TYPE_POOL_SHARE);
+      xdr.setLiquidityPoolID(liquidityPoolId.toXdr());
+    }
+    return xdr;
+  }
 
-      return xdr;
+  /**
+   * Creates a TrustLineAsset from its XDR representation.
+   *
+   * @param trustLineAsset The XDR representation of the TrustLineAsset.
+   * @return A new TrustLineAsset instance.
+   * @throws IllegalArgumentException if the asset type is unknown.
+   */
+  public static TrustLineAsset fromXdr(org.stellar.sdk.xdr.TrustLineAsset trustLineAsset) {
+    switch (trustLineAsset.getDiscriminant()) {
+      case ASSET_TYPE_NATIVE:
+        return new TrustLineAsset(new AssetTypeNative());
+      case ASSET_TYPE_CREDIT_ALPHANUM4:
+        return new TrustLineAsset(AssetTypeCreditAlphaNum4.fromXdr(trustLineAsset.getAlphaNum4()));
+      case ASSET_TYPE_CREDIT_ALPHANUM12:
+        return new TrustLineAsset(
+            AssetTypeCreditAlphaNum12.fromXdr(trustLineAsset.getAlphaNum12()));
+      case ASSET_TYPE_POOL_SHARE:
+        return new TrustLineAsset(LiquidityPoolID.fromXdr(trustLineAsset.getLiquidityPoolID()));
+      default:
+        throw new IllegalArgumentException(
+            "Unknown asset type " + trustLineAsset.getDiscriminant());
     }
   }
 }