+++ src/main/java/org/stellar/sdk/operations/LiquidityPoolDepositOperation.java
@@ -8,11 +8,10 @@ import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 import org.stellar.sdk.AccountConverter;
 import org.stellar.sdk.AssetAmount;
+import org.stellar.sdk.LiquidityPool;
 import org.stellar.sdk.LiquidityPoolID;
-import org.stellar.sdk.LiquidityPoolParameters;
 import org.stellar.sdk.Price;
 import org.stellar.sdk.xdr.LiquidityPoolDepositOp;
-import org.stellar.sdk.xdr.LiquidityPoolType;
 import org.stellar.sdk.xdr.Operation.OperationBody;
 import org.stellar.sdk.xdr.OperationType;
 
@@ -50,12 +49,7 @@ public class LiquidityPoolDepositOperation extends Operation {
     if (a.getAsset().compareTo(b.getAsset()) >= 0) {
       throw new IllegalArgumentException("AssetA must be < AssetB");
     }
-    this.liquidityPoolID =
-        new LiquidityPoolID(
-            LiquidityPoolType.LIQUIDITY_POOL_CONSTANT_PRODUCT,
-            a.getAsset(),
-            b.getAsset(),
-            LiquidityPoolParameters.FEE);
+    this.liquidityPoolID = new LiquidityPool(a.getAsset(), b.getAsset()).getLiquidityPoolId();
     this.maxAmountA = a.getAmount();
     this.maxAmountB = b.getAmount();
     this.minPrice = minPrice;