+++ src/test/java/org/stellar/sdk/operations/ChangeTrustOperationTest.java
@@ -4,11 +4,11 @@ import static junit.framework.TestCase.assertEquals;
 
 import org.junit.Assert;
 import org.junit.Test;
+import org.stellar.sdk.Asset;
 import org.stellar.sdk.AssetTypeCreditAlphaNum12;
 import org.stellar.sdk.AssetTypeCreditAlphaNum4;
 import org.stellar.sdk.ChangeTrustAsset;
-import org.stellar.sdk.LiquidityPoolParameters;
-import org.stellar.sdk.xdr.LiquidityPoolType;
+import org.stellar.sdk.LiquidityPool;
 
 public class ChangeTrustOperationTest {
   @Test
@@ -22,8 +22,8 @@ public class ChangeTrustOperationTest {
     // )
     // print(op.to_xdr_object().to_xdr())
     ChangeTrustAsset changeTrustAsset =
-        ChangeTrustAsset.createNonNativeAsset(
-            "USDC", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN");
+        new ChangeTrustAsset(
+            Asset.create("USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"));
     String limit = "922337203685.4775807";
     String source = "GASOCNHNNLYFNMDJYQ3XFMI7BYHIOCFW3GJEOWRPEGK2TDPGTG2E5EDW";
     ChangeTrustOperation op =
@@ -55,15 +55,13 @@ public class ChangeTrustOperationTest {
     //    source="GASOCNHNNLYFNMDJYQ3XFMI7BYHIOCFW3GJEOWRPEGK2TDPGTG2E5EDW",
     // )
     // print(op.to_xdr_object().to_xdr())
-    LiquidityPoolParameters liquidityPoolParameters =
-        LiquidityPoolParameters.create(
-            LiquidityPoolType.LIQUIDITY_POOL_CONSTANT_PRODUCT,
+    LiquidityPool liquidityPool =
+        new LiquidityPool(
             new AssetTypeCreditAlphaNum4(
                 "USD", "GCNY5OXYSY4FKHOPT2SPOQZAOEIGXB5LBYW3HVU3OWSTQITS65M5RCNY"),
             new AssetTypeCreditAlphaNum12(
-                "CATCOIN", "GDJVFDG5OCW5PYWHB64MGTHGFF57DRRJEDUEFDEL2SLNIOONHYJWHA3Z"),
-            LiquidityPoolParameters.FEE);
-    ChangeTrustAsset changeTrustAsset = ChangeTrustAsset.create(liquidityPoolParameters);
+                "CATCOIN", "GDJVFDG5OCW5PYWHB64MGTHGFF57DRRJEDUEFDEL2SLNIOONHYJWHA3Z"));
+    ChangeTrustAsset changeTrustAsset = new ChangeTrustAsset(liquidityPool);
     String limit = "922337203685.4775807";
     String source = "GASOCNHNNLYFNMDJYQ3XFMI7BYHIOCFW3GJEOWRPEGK2TDPGTG2E5EDW";
     ChangeTrustOperation op =
@@ -83,8 +81,8 @@ public class ChangeTrustOperationTest {
   @Test
   public void testFromXdrWithAsset() {
     ChangeTrustAsset changeTrustAsset =
-        ChangeTrustAsset.createNonNativeAsset(
-            "USDC", "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN");
+        new ChangeTrustAsset(
+            Asset.create("USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"));
     String limit = "922337203685.4775807";
     String source = "GASOCNHNNLYFNMDJYQ3XFMI7BYHIOCFW3GJEOWRPEGK2TDPGTG2E5EDW";
     ChangeTrustOperation op =
@@ -100,15 +98,13 @@ public class ChangeTrustOperationTest {
 
   @Test
   public void testFromXdrWithLiquidityPoolAsset() {
-    LiquidityPoolParameters liquidityPoolParameters =
-        LiquidityPoolParameters.create(
-            LiquidityPoolType.LIQUIDITY_POOL_CONSTANT_PRODUCT,
+    LiquidityPool liquidityPool =
+        new LiquidityPool(
             new AssetTypeCreditAlphaNum4(
                 "USD", "GCNY5OXYSY4FKHOPT2SPOQZAOEIGXB5LBYW3HVU3OWSTQITS65M5RCNY"),
             new AssetTypeCreditAlphaNum12(
-                "CATCOIN", "GDJVFDG5OCW5PYWHB64MGTHGFF57DRRJEDUEFDEL2SLNIOONHYJWHA3Z"),
-            LiquidityPoolParameters.FEE);
-    ChangeTrustAsset changeTrustAsset = ChangeTrustAsset.create(liquidityPoolParameters);
+                "CATCOIN", "GDJVFDG5OCW5PYWHB64MGTHGFF57DRRJEDUEFDEL2SLNIOONHYJWHA3Z"));
+    ChangeTrustAsset changeTrustAsset = new ChangeTrustAsset(liquidityPool);
     String limit = "922337203685.4775807";
     String source = "GASOCNHNNLYFNMDJYQ3XFMI7BYHIOCFW3GJEOWRPEGK2TDPGTG2E5EDW";
     ChangeTrustOperation op =