+++ src/test/java/org/stellar/sdk/responses/AssetDeserializerTest.java
@@ -4,13 +4,14 @@ import junit.framework.TestCase;
 import org.junit.Test;
 import org.stellar.sdk.Asset;
 import org.stellar.sdk.AssetTypeCreditAlphaNum;
+import org.stellar.sdk.xdr.AssetType;
 
 public class AssetDeserializerTest extends TestCase {
   @Test
   public void testDeserializeNative() {
     String json = "{\"asset_type\": \"native\"}";
     Asset asset = GsonSingleton.getInstance().fromJson(json, Asset.class);
-    assertEquals(asset.getType(), "native");
+    assertEquals(asset.getType(), AssetType.ASSET_TYPE_NATIVE);
   }
 
   @Test
@@ -22,7 +23,7 @@ public class AssetDeserializerTest extends TestCase {
             + "  \"asset_issuer\": \"GAREELUB43IRHWEASCFBLKHURCGMHE5IF6XSE7EXDLACYHGRHM43RFOX\"\n"
             + "}";
     Asset asset = GsonSingleton.getInstance().fromJson(json, Asset.class);
-    assertEquals(asset.getType(), "credit_alphanum4");
+    assertEquals(asset.getType(), AssetType.ASSET_TYPE_CREDIT_ALPHANUM4);
     AssetTypeCreditAlphaNum creditAsset = (AssetTypeCreditAlphaNum) asset;
     assertEquals(creditAsset.getCode(), "CNY");
     assertEquals(
@@ -33,14 +34,14 @@ public class AssetDeserializerTest extends TestCase {
   public void testDeserializeCanonicalStringNative() {
     String json = "\"native\"";
     Asset asset = GsonSingleton.getInstance().fromJson(json, Asset.class);
-    assertEquals(asset.getType(), "native");
+    assertEquals(asset.getType(), AssetType.ASSET_TYPE_NATIVE);
   }
 
   @Test
   public void testDeserializeCanonicalStringCredit() {
     String json = "\"CNY:GAREELUB43IRHWEASCFBLKHURCGMHE5IF6XSE7EXDLACYHGRHM43RFOX\"";
     Asset asset = GsonSingleton.getInstance().fromJson(json, Asset.class);
-    assertEquals(asset.getType(), "credit_alphanum4");
+    assertEquals(asset.getType(), AssetType.ASSET_TYPE_CREDIT_ALPHANUM4);
     AssetTypeCreditAlphaNum creditAsset = (AssetTypeCreditAlphaNum) asset;
     assertEquals(creditAsset.getCode(), "CNY");
     assertEquals(